feat(env-loader): support envVar preload priority and dotenv options
Allow resolving preferred .env.<value> via envVar, forward dotenv config (except path), and consolidate Env into index.ts.
